fix(firewall): harden route-identity persistence, seed lifecycle, and observability
Review fixes for the sticky route-identity rollout:

- IdentityAllocator: retry persistence on next sync when a write fails
  (needsPersist flag) instead of silently dropping the table; treat an
  out-of-range persisted cursor on a non-empty table as a corrupt-table
  startup gate; guard zero-value use; delete the test-only SyncFromRules
  wrapper.
- SyncRoutes: surface failed orphan-seed deletes at Warn
  (event=dns_seed_orphan_delete_failed), join them into the returned
  error, and retain the IP for retry — seeds are GC-immortal, so the
  delete is their only removal path (extracted reconcileSeeds helper).
- dnsbpf: skip dns_cache writes when the seed-precedence lookup fails
  with anything other than ErrKeyNotExist; reject duplicate dnsbpf
  directives per server block (plugin.ErrOnce); new setup parse-error
  tests.
- Resolver-miss observability: RoutesFromRules and GenerateCorefile
  return missed dsts; handler and stack emit
  event=identity_resolver_miss; NewStack warns on a nil resolver
  (event=identity_resolver_unset).
- Type safety: new ebpf.RouteIdentity named type adopted across
  firewall → dnsbpf → netlogger so identity/IP/TTL transposition no
  longer compiles; zero sentinel exposed as RouteIdentity.IsNone.
- Break-glass sync-routes rejects routes with zero identity/dstPort.
- Tests: persist-failure retry, corrupt-table gate, release→restart
  no-reuse round-trip, fail-closed resolver-miss paths.
- Docs/comments: correct fail-closed layer attribution (Envoy vs BPF
  hook), seed-protection mechanism (DNSSourceSeed tag), residual
  domain-hash vocabulary, and history clauses.

+++ controlplane/firewall/ebpf/netlogger/event_test.go
@@ -42,7 +42,7 @@ func TestParseEvent_RoundTrip(t *testing.T) {
 	if got.BPFTsNs != in.TsNs {
 		t.Errorf("BPFTsNs = %d; want %d", got.BPFTsNs, in.TsNs)
 	}
-	if got.Identity != in.Identity {
+	if uint32(got.Identity) != in.Identity {
 		t.Errorf("Identity = %x; want %x", got.Identity, in.Identity)
 	}
 	want := netip.AddrFrom4([4]byte{203, 0, 113, 7})
